Add files via upload
here is the code for this project:-
I have created a random move based scramble generator for the 3x3 rubiks cube to be used with a stackmat timer. Once you scramble up your cube take the stackmat timer and time your solve and enter it . This program also calculates the Ao5. This can be used if you want to practice for an upcoming comp. I have used python  with gui for the project